requirements.txt

pylast~=4.0.0
python-dotenv~=0.15.0

flask-sqlalchemy
flask~=1.1.2

SqlAlchemy~=1.3.20
sqlalchemy_utils
cryptography
pymysql
pytest

zope.Interface

Werkzeug~=1.0.1
spotipy~=2.16.1